updated tests for new dir test structure

+++ tests/conftest.py
@@ -312,7 +312,6 @@ def application_user(db, oauth_client):
         db=db,
         data={
             "username": unique_username,
-            "password": "test_password",
             "email_address": f"{unique_username}@ethyca.com",
             "first_name": "Test",
             "last_name": "User",
@@ -325,32 +324,35 @@ def application_user(db, oauth_client):
 
 @pytest.fixture
 def user(db):
-    user = FidesUser.create(
+    created, user = FidesUser.get_or_create(
         db=db,
         data={
             "username": "test_fidesops_user",
-            "password": "TESTdcnG@wzJeu0&%3Qe2fGo7",
             "email_address": "fides.user@ethyca.com",
         },
     )
-    client = ClientDetail(
-        hashed_secret="thisisatest",
-        salt="thisisstillatest",
-        roles=[APPROVER],
-        scopes=[],
-        user_id=user.id,
-    )
+    if created:
+        client = ClientDetail(
+            hashed_secret="thisisatest",
+            salt="thisisstillatest",
+            roles=[APPROVER],
+            scopes=[],
+            user_id=user.id,
+        )
 
-    FidesUserPermissions.create(db=db, data={"user_id": user.id, "roles": [APPROVER]})
+        FidesUserPermissions.create(
+            db=db, data={"user_id": user.id, "roles": [APPROVER]}
+        )
 
-    db.add(client)
-    db.commit()
-    db.refresh(client)
+        db.add(client)
+        db.commit()
+        db.refresh(client)
     yield user
-    try:
-        client.delete(db)
-    except ObjectDeletedError:
-        pass
+    if created:
+        try:
+            client.delete(db)
+        except ObjectDeletedError:
+            pass
 
 
 @pytest.fixture
